adjust

+++ jdk8-useful-utils/src/main/java/_collections/set/StackDemo.java
@@ -7,15 +7,13 @@ public class StackDemo {
     static void showPush(java.util.Stack<Integer> st, int a) {
         st.push(new Integer(a));
 
-        System.out.println("push(" + a + ") -> " +
-                "stack: " + st);
+        System.out.println("push(" + a + ") -> " + "stack: " + st);
     }
 
     static void showPop(java.util.Stack<Integer> st) {
-        Integer a = (Integer) st.pop();
+        Integer a = st.pop();
 
-        System.out.println("pop(" + a + ") -> " +
-                "stack: " + st);
+        System.out.println("pop(" + a + ") -> " + "stack: " + st);
     }
 
     public static void main(String args[]) {
